Refactor and remove unused code
The problem with the nonce source has been finally solved. Each session maintains a separate collection of nonce values returned by the ACME server. Each request uses and removes a nonce value from this collection. If the collection is empty by the time a new request is to be issued, a separate request is made to the ACME server to return a new nonce value.